Actually, *don't* use session state for CSRF/nonce data
Using session state means that we are forced to mark the session state
as SameSite::Lax, which opens us up to CSRF attacks on GET requests.

*However,* even that doesn't work as-is, because the Tide session cookie
is still marked (correctly!) as SameSite::Strict and so our attempt to
mark the session as "authenticated" fails when we do the final redirect
to the landing URL. The safest fix is to use an HTML/client-side redirect
to ensure that the (new) session cookie is persisted at the final step,
and thus the browser ends up with an authenticated session.