(back to using) Session state for the OCID CSRF/Nonce
This requires the session cookie to use `SameSite::Lax`, but I think that
is appropriate given what we are trying to do here with the OIDC middleware.
First, using this middleware at all means that you are in a bit of a
complicated environment, so I think it is reasonable to ask that folks
in that environment avoid mutating state with a GET request (which is
*the only* advantage to Strict vs. Lax; because we are protected with
all HTTP-level mutation operations through `tide-csrf` and so we do not
need Strict there). Second, retaining the same session across the auth
flow means that you can accumulate (session) state *prior* to auth, thus
allowing things like guest access (with state!) before sign in.